feat(mcp): extract registry asset helpers into registry-asset-lib
Isolate copyable asset shaping and install complexity scoring for focused
testing without changing entry.asset or toolbox behavior.
