+++ packages/mcp/src/registry.js
@@ -96,6 +96,11 @@ import {
   entryMatchesTrustFilters,
 } from "./registry-filter-lib.js";
 import { projectEntryBody } from "./registry-excerpt-lib.js";
+import {
+  categoryPrimaryAsset,
+  contentAsset,
+  entryInstallComplexity,
+} from "./registry-asset-lib.js";
 
 export {
   LOCAL_DRAFT_TOOL_NAMES,
@@ -418,85 +423,6 @@ function entryTrustSignalCoverage(entry) {
     present: presentOrdered,
     missing,
   };
-}
-
-function contentAsset(type, label, content, format = "markdown") {
-  const text =
-    content && typeof content === "object"
-      ? JSON.stringify(content, null, 2)
-      : String(content || "").trim();
-  if (!text) return null;
-  return {
-    type,
-    label,
-    format,
-    content: text,
-    length: text.length,
-  };
-}
-
-function categoryPrimaryAsset(entry) {
-  const assets = [
-    contentAsset(
-      "full_content",
-      "Full usable entry content",
-      entry.fullCopyableContent || entry.copySnippet || entry.body,
-    ),
-    contentAsset(
-      "install_command",
-      "Install command",
-      entry.installCommand,
-      "shell",
-    ),
-    contentAsset(
-      "config_snippet",
-      "Configuration snippet",
-      entry.configSnippet,
-      "text",
-    ),
-    contentAsset("script", "Script body", entry.scriptBody, "text"),
-    contentAsset(
-      "command_syntax",
-      "Command syntax",
-      entry.commandSyntax,
-      "text",
-    ),
-    contentAsset("usage", "Usage snippet", entry.usageSnippet, "markdown"),
-    contentAsset("items", "Collection items", entry.items, "json"),
-  ].filter(Boolean);
-
-  const preferredByCategory = {
-    agents: ["full_content", "usage"],
-    rules: ["full_content", "script", "usage"],
-    hooks: ["config_snippet", "script", "install_command", "usage"],
-    mcp: ["config_snippet", "install_command", "usage"],
-    skills: ["install_command", "full_content", "usage"],
-    statuslines: ["config_snippet", "script", "full_content", "usage"],
-    commands: ["command_syntax", "install_command", "full_content", "usage"],
-    collections: ["items", "full_content", "usage"],
-    guides: ["full_content", "usage"],
-  };
-  const preferred = preferredByCategory[entry.category] || ["full_content"];
-  return (
-    preferred
-      .map((type) => assets.find((asset) => asset.type === type))
-      .find(Boolean) ||
-    assets[0] ||
-    null
-  );
-}
-
-function entryInstallComplexity(entry) {
-  const pieces = [
-    entry.installCommand,
-    entry.configSnippet,
-    entry.downloadUrl,
-    entry.prerequisites,
-  ].filter((value) => String(value || "").trim());
-  if (pieces.length >= 3) return "higher";
-  if (pieces.length === 2) return "medium";
-  if (pieces.length === 1) return "low";
-  return "unknown";
 }
 
 async function readEntry(category, slug, options = {}) {
